Config propiedad Cuil y EncriptPassword

+++ DTOs/UsuarioLoginDto.cs
@@ -6,7 +6,7 @@
         public string Nombre { get; set; }
         public int RoleId { get; set; }
         public int? Dni { get; set; }
-        public int? Cuil { get; set; }
+        public long? Cuil { get; set; }
         public string Token { get; set; }
     }
 }
